# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CSMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers with pagination and filters
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          schema:
            type: string
        - name: branch_id
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: governorate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomersListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerItem"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Customer retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerItem"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerItem"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Customer deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /complaints/{id}/logs:
    get:
      operationId: getComplaintLogs
      tags: [complaints]
      summary: Get complaint activity logs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Complaint logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintLogsListResponse"
...
  /follow-ups:
    get:
      operationId: listFollowUps
      tags: [follow-ups]
      summary: List follow-ups
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Follow-ups list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUpsListResponse"
    post:
      operationId: createFollowUp
      tags: [follow-ups]
      summary: Create a follow-up
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowUpRequest"
      responses:
        "201":
          description: Follow-up created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUpItem"
...
  /import-logs:
    get:
      operationId: listImportLogs
      tags: [import-logs]
      summary: List import logs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Import logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportLogsListResponse"
  /import-logs/{id}:
    get:
      operationId: getImportLog
      tags: [import-logs]
      summary: Get import log detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Import log detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportLogItem"
...
  /analytics/customers:
    get:
      operationId: getCustomerAnalytics
      tags: [analytics]
      summary: Get customer analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAnalytics"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
...
    CustomerItem:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        phone:
          type: string
        type:
          type: string
        governorate:
          type: string
        branch_id:
          type: integer
        branch_name:
          type: string
        address:
          type: string
        invoice_count:
          type: integer
        last_invoice_date:
          type: string
      required:
        - id
        - code
        - name
        - phone
        - type
        - governorate
        - branch_id
    CustomersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CustomerItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    CreateCustomerRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        phone:
          type: string
        type:
          type: string
        governorate:
          type: string
        branch_id:
          type: integer
        address:
          type: string
      required:
        - name
        - phone
        - type
        - governorate
        - branch_id
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        type:
          type: string
        governorate:
          type: string
        branch_id:
          type: integer
        address:
          type: string
...
    ComplaintLogItem:
      type: object
      properties:
        id:
          type: integer
        complaint_id:
          type: integer
        action:
          type: string
        user_id:
          type: integer
        user_name:
          type: string
        note:
          type: string
        timestamp:
          type: string
      required:
        - id
        - complaint_id
        - action
        - user_id
        - timestamp
    ComplaintLogsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ComplaintLogItem"
      required:
        - data
...
    FollowUpItem:
      type: object
      properties:
        id:
          type: integer
        invoice_id:
          type: integer
        invoice_number:
          type: string
        customer_name:
          type: string
        branch_name:
          type: string
        assigned_user_id:
          type: integer
        assigned_user_name:
          type: string
        notes:
          type: object
        status:
          type: string
        created_at:
          type: string
      required:
        - id
        - invoice_id
        - assigned_user_id
        - status
    FollowUpsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FollowUpItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    CreateFollowUpRequest:
      type: object
      properties:
        invoice_ids:
          type: array
          items:
            type: integer
        notes:
          type: object
        assigned_user_id:
          type: integer
      required:
        - invoice_ids
        - assigned_user_id
...
    ImportLogItem:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
        file_name:
          type: string
        total_rows:
          type: integer
        added_customers:
          type: integer
        updated_customers:
          type: integer
        added_invoices:
          type: integer
        duplicate_invoices:
          type: integer
        errors:
          type: array
          items:
            type: object
        warnings:
          type: array
          items:
            type: object
        imported_at:
          type: string
      required:
        - id
        - user_id
        - file_name
        - total_rows
    ImportLogsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ImportLogItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...
    CustomerAnalytics:
      type: object
      properties:
        top_by_purchase:
          type: array
          items:
            type: object
        by_complaints:
          type: array
          items:
            type: object
        branch_changes:
          type: array
          items:
            type: object
      required:
        - top_by_purchase
...